requirements.txt
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
coverage==7.11.0
cryptography==46.0.3
idna==3.11
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
pycparser==2.23
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.0
pytest==8.4.2
pytest-cov==7.0.0
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
flake8==7.1.1
pytest-mock==3.12.0